requirements.txt
Django==2.2.10
django-unixtimestampfield==0.3.9
mysqlclient==1.4.6
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
